refactor: extract EvidenceFileOperations and EvidenceJsonWriter from SaveOperationCompletionEvidence
Reduce SaveOperationCompletionEvidence.java from 1009 to 429 lines by
extracting two package-private helper classes:

- EvidenceFileOperations (141 lines): RegularFileState record, path traversal
  guards, symlink protection, file inspection, and canonical directory helpers.

- EvidenceJsonWriter (587 lines): all JSON string builders for result,
  dialog control, invocation proof, and save proof artifacts. Includes
  SaveProofSnapshot record for atomic state capture from SaveProofEvidence.

Inner classes SaveProofEvidence and InvocationTrigger remain nested to
preserve 30+ qualified references in tests and AbstractSaveOperation.
No API changes, no visibility escalation, no test modifications.

Closes #561
